requirements.txt
asgiref==3.9.2
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.0.3
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==21.2.0
idna==3.10
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
